feat: move runner and nanocontract core to hathorlib
